Create confignetwork.sh
Added a script that configures the network interfaces. If the rover is unable to connect to a wireless network it is laggy and does not response well, as 100% of the cpu every 15 seconds is spent finding a network. This script checks for internet access, and if it does not find it turns off the UCI parameter that enables connecting to an existing network. This script is run on boot